requirements.txt
requests
Flask==2.0.2
Jinja2==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.2
openai
PyPDF2